# swagger/swagger.yaml
swagger: '2.0'
info:
  title: ShoppingApp
  description: Neosoft Shopping App
  version: 1.0.0
host: 'localhost:3000'
schemes:
  - http
paths:
  /api/categories:
    post:
      tags:
        - Category
      summary: Add categories
      description: Add categories
      operationId: addCategories
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "Categories" model
          required: true
          schema:
            $ref: '#/definitions/Categories'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    get:
      tags:
        - Category
      summary: Get categories
      description: Get categories
      operationId: getCategories
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security:  
      - authorization: []
  /api/products:
    post:
      tags:
        - Product
      summary: Add product
      description: Add product
      operationId: addProduct
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "Product" model
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    get:
      tags:
        - Product
      summary: Get Products
      description: Get Products
      operationId: getProducts
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security:  
      - authorization: []
  /api/categories/{categoryId}:
    get:
      tags:
        - Category
      summary: Get category by id
      description: Get category by id
      operationId: getCategoryById
      produces:
        - application/json
      parameters:
      - name: "categoryId"
        in: "path"
        description: "_id of category to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
    delete:
      tags:
        - Category
      summary: Delete categories
      description: Delete categories
      operationId: deleteCategories
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: "categoryId"
        in: "path"
        description: "_id of category to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    put:
      tags:
        - Category
      summary: Update categories
      description: Update categories
      operationId: updateCategories
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "categoryId"
          in: "path"
          description: "_id of category to return"
          required: true
          type: "string"
        - name: body
          in: body
          description: the "Categories" model
          required: true
          schema:
            $ref: '#/definitions/Categories'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/{productId}:
    delete:
      tags:
        - Product
      summary: Delete Product
      description: Delete Product
      operationId: deleteProduct
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: "productId"
        in: "path"
        description: "_id of productId to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    put:
      tags:
        - Product
      summary: Update Product
      description: Update Product
      operationId: updateProduct
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "productId"
          in: "path"
          description: "_id of category to return"
          required: true
          type: "string"
        - name: body
          in: body
          description: the "Product" model
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/home-list:
    get:
      tags:
        - Product
      summary: Get product list under categories
      description: Get product list under categories
      operationId: getProductListUnderCategories
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/products/addProductReview:
    post:
      tags:
        - Product
      summary: Add Product Review and Rating
      description: Add Product Review and Rating
      operationId: addProductReview
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "ProductReview" model
          required: true
          schema:
            $ref: '#/definitions/ProductReview'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/{productId}/similarProducts:
    get:
      tags:
        - Product
      summary: Get Products of similar specification 
      description: Get Products of similar specification 
      operationId: getSimilarProducts
      produces:
        - application/json
      parameters:
      - name: "productId"
        in: "path"
        description: "_id of product to return"
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page Number"
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "Record limit"
        required: false
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/getProductReviewById/{productId}:
    get:
      tags:
        - Product
      summary: Get Product Review and Rating
      description: Get Product Review and Rating
      operationId: getProductReviewById
      produces:
        - application/json
      parameters:
      - name: "productId"
        in: "path"
        description: "_id of product to return"
        required: true
        type: "string"
      - name: "page"
        in: "query"
        description: "Page Number"
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "Record limit"
        required: false
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/{productId}/editProductReview/{reviewId}:
    put:
      tags:
        - Product
      summary: Edit Product Review and Rating
      description: Edit Product Review and Rating
      operationId: editProductReview
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "productId"
          in: "path"
          description: "_id of product to return"
          required: true
          type: "string"
        - name: "reviewId"
          in: "path"
          description: "_id of review to return"
          required: true
          type: "string"
        - name: body
          in: body
          description: the "EditProductReview" model
          required: true
          schema:
            $ref: '#/definitions/EditProductReview'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/products/{productId}/deleteProductReview/{reviewId}:
    delete:
      tags:
        - Product
      summary: Delete Product Review and Rating
      description: Delete Product Review and Rating
      operationId: deleteProductReview
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "productId"
          in: "path"
          description: "_id of product to return"
          required: true
          type: "string"
        - name: "reviewId"
          in: "path"
          description: "_id of review to return"
          required: true
          type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/users/{userId}:
    get:
      tags:
        - User
      summary: Get user by Id
      description: Get user by Id
      operationId: getUserById
      produces:
        - application/json
      parameters:
      - name: "userId"
        in: "path"
        description: "_id of userId to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    put:
      tags:
        - User
      summary: Update user by Id
      description: Update user by Id
      operationId: updateUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "userId"
          in: "path"
          description: "_id of userId to return"
          required: true
          type: "string"
        - name: body
          in: body
          description: the "UpdateUser" model
          required: true
          schema:
            $ref: '#/definitions/UpdateUser'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    delete:
      tags:
        - User
      summary: Delete User
      description: Delete User
      operationId: deleteUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: "userId"
        in: "path"
        description: "_id of userId to return"
        required: true
        type: "string"
      - name: body
        in: body
        description: the "AuthUser" model
        required: true
        schema:
          $ref: '#/definitions/AuthUser'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/users:
    get:
      tags:
        - User
      summary: Users
      description: Users
      operationId: users
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User signin
      description: User signin
      operationId: userSignin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "AuthUser" model
          required: true
          schema:
            $ref: '#/definitions/AuthUser'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/auth/sign-up:
    post:
      tags:
        - Auth
      summary: User signup
      description: User signup
      operationId: userSignup
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "User" model
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: User ForgotPassword
      description: User ForgotPassword
      operationId: userForgotPassword
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "User" model
          required: true
          schema:
            $ref: '#/definitions/ForgotPassword'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/carts:
    post:
      tags:
        - Cart
      summary: Add product into Cart
      description: Add product into Cart
      operationId: addProductIntoCart
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "Cart" model
          required: true
          schema:
            $ref: '#/definitions/Cart'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    get:
      tags:
        - Cart
      summary: Get Carts
      description: Get Carts
      operationId: getCarts
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/carts/{cartId}:
    put:
      tags:
        - Cart
      summary: Update Cart
      description: Update Cart
      operationId: updateCart
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: "cartId"
          in: "path"
          description: "_id of cart to return"
          required: true
          type: "string"
        - name: body
          in: body
          description: the "Cart" model
          required: true
          schema:
            $ref: '#/definitions/Cart'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
    delete:
      tags:
        - Cart
      summary: Delete cart by id
      description: Delete cart by id
      operationId: deleteCartById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: "cartId"
        in: "path"
        description: "_id of cart to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: 
      - authorization: []
  /api/categories/dashboard-products:
    get:
      tags:
        - Site
      summary: Get Categories Dashboard Products
      description: Get Categories Dashboard Products
      operationId: getCategoriesDashboardProducts
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/categories/{categoryId}/brand-count:
    get:
      tags:
        - Site
      summary: Get category brand count
      description: Get category brand count
      operationId: getBrandCountByCategoryId
      produces:
        - application/json
      parameters:
      - name: "categoryId"
        in: "path"
        description: "_id of category to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/products/byCategoryId/{categoryId}:
    get:
      tags:
        - Site
      summary: Get products by category
      description: Get products by category
      operationId: getProductsByCategoryId
      produces:
        - application/json
      parameters:
      - name: "categoryId"
        in: "path"
        description: "_id of category to return"
        required: true
        type: "string"
      - name: "pageNo"
        in: "query"
        description: "Page Number"
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "Record limit"
        required: false
        type: "string"
      - name: "brandId"
        in: "query"
        description: "_id of brand to return"
        required: false
        type: "string"
      - name: "range"
        in: "query"
        description: "price range as 500-1000"
        required: false
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
definitions:
  Product:
    description: Model for Product
    required:
      - brand
    properties:
      brand:
        type: string
        x-example: Motorola
      category_id:
        type: string
        x-example: 5d2d9bf1b3af78720a1f0a54
      name:
        type: string
        x-example: Moto G
      price:
        type: string
        x-example: '8999'
  ProductReview:
    description: Model for ProductReview
    required:
      - productId
    properties:
      productId:
        type: string
        x-example: 5d2b6201fc7a944e5f080cc3
      rating:
        type: number
        x-example: 5
      comment:
        type: string
        x-example: Good Product
  EditProductReview:
    description: Model for ProductReview
    required:
      - rating
      - comment
    properties:
      rating:
        type: number
        x-example: 5
      comment:
        type: string
        x-example: Good Product
  User:
    description: Model for User
    required:
      - confirm_password
    properties:
      confirm_password:
        type: string
        x-example: jitendra123
      email:
        type: string
        x-example: jitendra.kumar@neosofttech.com
      first_name:
        type: string
        x-example: Jitendra
      gender:
        type: string
        x-example: Male
      last_name:
        type: string
        x-example: Kumar
      password:
        type: string
        x-example: jitendra123
  UpdateUser:
    description: Model for UpdateUser
    required:
      - email
    properties:
      email:
        type: string
        x-example: jitendra.kumar@neosofttech.com
      password:
        type: string
        x-example: jitendra123
      first_name:
        type: string
        x-example: Jitendra
      last_name:
        type: string
        x-example: Kumar
  AuthUser:
    description: Model for AuthUser
    required:
      - email
      - password
    properties:
      email:
        type: string
        x-example: jitendra.kumar@neosofttech.com
      password:
        type: string
        x-example: jitendra123
  ForgotPassword:
    description: Model for ForgotPassword
    required:
      - email
    properties:
      email:
        type: string
        x-example: jitendra.kumar@neosofttech.com
  Cart:
    description: Model for Cart
    required:
      - product_id
    properties:
      product_id:
        type: string
        x-example: 5d2d9cf3b3af78720a1f0a57
      quantity:
        type: integer
        format: int32
        x-example: '1'
  Categories:
    description: Model for Categories
    required:
      - description
    properties:
      description:
        type: string
        x-example: mobile phone
      name:
        type: string
        x-example: mobile
securityDefinitions: 
  authorization:
    type: "apiKey"
    name: "authorization"
    in: "header"
security: []
tags:
  - name: User
    description: Operations about User
  - name: Auth
    description: Operations about Auth
  - name: Product
    description: Operations about Product
  - name: Category
    description: Operations about Category
  - name: Site
    description: Operations about Site